update currency user interface to new design system
