+++ core/ui/navigation/routes.tsx
@@ -1,16 +1,16 @@
+import { ReshareVaultPage } from '@core/ui/mpc/keygen/reshare/ReshareVaultPage'
+import { CorePath } from '@core/ui/navigation'
+import { CurrencyPage } from '@core/ui/preferences/currency'
+import { ActiveVaultGuard } from '@core/ui/vault/ActiveVaultGuard'
 import { ReactNode } from 'react'
 
-import { ReshareVaultPage } from '../mpc/keygen/reshare/ReshareVaultPage'
-import { CurrencyPage } from '../preferences/currency/CurrencyPage'
-import { ActiveVaultGuard } from '../vault/ActiveVaultGuard'
-
-type SharedPaths = 'reshareVault' | 'currencySettings'
+type SharedPaths = Extract<CorePath, 'currencySettings' | 'reshareVault'>
 
 export const sharedRoutes: Record<SharedPaths, ReactNode> = {
+  currencySettings: <CurrencyPage />,
   reshareVault: (
     <ActiveVaultGuard>
       <ReshareVaultPage />
     </ActiveVaultGuard>
   ),
-  currencySettings: <CurrencyPage />,
 }
