+++ core/ui/preferences/currency/index.tsx
@@ -3,21 +3,20 @@ import {
   useFiatCurrency,
   useSetFiatCurrencyMutation,
 } from '@core/ui/storage/fiatCurrency'
-import { CheckIcon } from '@lib/ui/icons/CheckIcon'
 import { VStack } from '@lib/ui/layout/Stack'
+import { List } from '@lib/ui/list'
+import { ListItem } from '@lib/ui/list/item'
+import { ListItemTag } from '@lib/ui/list/item/tag'
 import { PageContent } from '@lib/ui/page/PageContent'
 import { PageHeader } from '@lib/ui/page/PageHeader'
 import { PageHeaderBackButton } from '@lib/ui/page/PageHeaderBackButton'
 import { PageHeaderTitle } from '@lib/ui/page/PageHeaderTitle'
-import { Text } from '@lib/ui/text'
 import { useTranslation } from 'react-i18next'
-
-import { CurrencyBox, CurrencyButton } from './CurrencyPage.styles'
 
 export const CurrencyPage = () => {
   const { t } = useTranslation()
-  const fiatCurrency = useFiatCurrency()
-  const { mutate: setFiatCurrency } = useSetFiatCurrencyMutation()
+  const currencyValue = useFiatCurrency()
+  const currencyMutation = useSetFiatCurrencyMutation()
 
   return (
     <VStack fullHeight>
@@ -26,17 +25,22 @@ export const CurrencyPage = () => {
         title={<PageHeaderTitle>{t('currency')}</PageHeaderTitle>}
         hasBorder
       />
-      <PageContent gap={16} flexGrow scrollable>
-        {fiatCurrencies.map((fiat, index) => (
-          <CurrencyButton key={index} onClick={() => setFiatCurrency(fiat)}>
-            <CurrencyBox>
-              <Text color="contrast" size={16} weight="600">
-                {fiat.toUpperCase()}
-              </Text>
-            </CurrencyBox>
-            {fiat === fiatCurrency && <CheckIcon />}
-          </CurrencyButton>
-        ))}
+      <PageContent flexGrow scrollable>
+        <List>
+          {fiatCurrencies.map(key => (
+            <ListItem
+              extra={
+                key === currencyValue && (
+                  <ListItemTag status="success" title={t('active')} />
+                )
+              }
+              key={key}
+              onClick={() => currencyMutation.mutate(key)}
+              title={key.toUpperCase()}
+              hoverable
+            />
+          ))}
+        </List>
       </PageContent>
     </VStack>
   )
